Completed all printer functions with the following changes:
- Added hex number printing
- Using unsigned int for len instead of size_t as printf returns size as
int instead of size_t
- Removed all unneeded dependencies. Printer functions no longer depends
on other files.

+++ printers/cs.c
@@ -6,7 +6,7 @@
 /*   By: htsang <htsang@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/03 17:24:36 by htsang            #+#    #+#             */
-/*   Updated: 2022/11/05 19:53:38 by htsang           ###   ########.fr       */
+/*   Updated: 2022/11/06 13:26:18 by htsang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	print_c(char c)
 	printer(&c, 1);
 }
 
-void	print_s(char *str, size_t len)
+void	print_s(char *str, unsigned int len)
 {
 	printer(str, len);
 }
