+++ printers/printer.c
@@ -6,13 +6,13 @@
 /*   By: htsang <htsang@student.42heilbronn.de>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/03 00:25:15 by htsang            #+#    #+#             */
-/*   Updated: 2022/11/05 19:25:17 by htsang           ###   ########.fr       */
+/*   Updated: 2022/11/06 13:26:01 by htsang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <unistd.h>
 
-int	*printer(char *str, size_t len)
+int	*printer(char *str, unsigned int len)
 {
 	static int	size = 0;
